+++ src/testmodule/weightingFormula.cpp
@@ -227,7 +227,7 @@ WeightingFunctionContextFormula::WeightingFunctionContextFormula(
 void WeightingFunctionContextFormula::addWeightingFeature(
 		const std::string& name_,
 		PostingIteratorInterface* itr_,
-		float weight_,
+		double weight_,
 		const TermStatistics& stats_)
 {
 	try
