+++ src/testmodule/weightingtest.cpp
@@ -21,7 +21,7 @@ namespace test {
 
 void WeightingFunctionContextTest::addWeightingFeature( const std::string& name,
 	strus::PostingIteratorInterface* itr,
-	float weight,
+	double weight,
 	const strus::TermStatistics& stats )
 {
 	try {
